Add config parser/validator to module
